deduplicate ExecuteWorkflowOptions / IWorkflowExecutionDataProcess
