+++ packages/cli/src/__tests__/active-executions.test.ts
@@ -1,9 +1,12 @@
 import { ActiveExecutions } from '@/active-executions';
 import PCancelable from 'p-cancelable';
 import { v4 as uuid } from 'uuid';
-import type { IExecuteResponsePromiseData, IRun } from 'n8n-workflow';
+import type {
+	IExecuteResponsePromiseData,
+	IRun,
+	IWorkflowExecutionDataProcess,
+} from 'n8n-workflow';
 import { createDeferredPromise } from 'n8n-workflow';
-import type { IWorkflowExecutionDataProcess } from '@/interfaces';
 import type { ExecutionRepository } from '@/databases/repositories/execution.repository';
 import { mock } from 'jest-mock-extended';
 import { ConcurrencyControlService } from '@/concurrency/concurrency-control.service';
