+++ packages/cli/src/commands/execute-batch.ts
@@ -3,7 +3,7 @@ import { Container } from 'typedi';
 import { Flags } from '@oclif/core';
 import fs from 'fs';
 import os from 'os';
-import type { IRun, ITaskData } from 'n8n-workflow';
+import type { IRun, ITaskData, IWorkflowExecutionDataProcess } from 'n8n-workflow';
 import { ApplicationError, jsonParse } from 'n8n-workflow';
 import { sep } from 'path';
 import { diff } from 'json-diff';
@@ -11,7 +11,7 @@ import pick from 'lodash/pick';
 
 import { ActiveExecutions } from '@/active-executions';
 import { WorkflowRunner } from '@/workflow-runner';
-import type { IWorkflowDb, IWorkflowExecutionDataProcess } from '@/interfaces';
+import type { IWorkflowDb } from '@/interfaces';
 import type { User } from '@/databases/entities/user';
 import { WorkflowRepository } from '@/databases/repositories/workflow.repository';
 import { OwnershipService } from '@/services/ownership.service';
