+++ packages/cli/src/commands/execute.ts
@@ -1,11 +1,10 @@
 import { Container } from 'typedi';
 import { Flags } from '@oclif/core';
-import type { IWorkflowBase } from 'n8n-workflow';
+import type { IWorkflowBase, IWorkflowExecutionDataProcess } from 'n8n-workflow';
 import { ApplicationError, ExecutionBaseError } from 'n8n-workflow';
 
 import { ActiveExecutions } from '@/active-executions';
 import { WorkflowRunner } from '@/workflow-runner';
-import type { IWorkflowExecutionDataProcess } from '@/interfaces';
 import { findCliWorkflowStart, isWorkflowIdValid } from '@/utils';
 import { BaseCommand } from './base-command';
