+++ packages/cli/src/wait-tracker.ts
@@ -1,6 +1,9 @@
-import { ApplicationError, ErrorReporterProxy as ErrorReporter } from 'n8n-workflow';
+import {
+	ApplicationError,
+	ErrorReporterProxy as ErrorReporter,
+	type IWorkflowExecutionDataProcess,
+} from 'n8n-workflow';
 import { Service } from 'typedi';
-import type { IWorkflowExecutionDataProcess } from '@/interfaces';
 import { WorkflowRunner } from '@/workflow-runner';
 import { ExecutionRepository } from '@/databases/repositories/execution.repository';
 import { OwnershipService } from '@/services/ownership.service';
